Updated the backend to support at max one reaction per user for each message     - database.go: made the comments unique in the comments table and created a new error for miltiple comments.     - db-message.go: added a check in the createComment function.     - messages.go: added the error handler for duplicate comments.

+++ service/api/messages.go
@@ -284,6 +284,13 @@ func (rt *_router) commentMessage(w http.ResponseWriter, r *http.Request, ps htt
 			return
 		}
 
+		if errors.Is(err, database.ErrUserAlreadyReacted) {
+			writeJSON(w, http.StatusConflict, errorResponse{
+				Message: "you already reacted to this message",
+			})
+			return
+		}
+
 		ctx.Logger.WithError(err).Error("cannot create comment")
 		writeJSON(w, http.StatusInternalServerError, errorResponse{
 			Message: "internal server error",
